+++ docs/changelog.md
@@ -14,6 +14,9 @@ Requires libvips v8.6.1.
   [#1290](https://github.com/lovell/sharp/pull/1290)
   [@sylvaindumont](https://github.com/sylvaindumont)
 
+* Ensure vendor platform mismatch throws error at install time.
+  [#1303](https://github.com/lovell/sharp/issues/1303)
+
 #### v0.20.5 - 27<sup>th</sup> June 2018
 
 * Expose libjpeg optimize_coding flag.