+++ docs/changelog.md
@@ -26,6 +26,9 @@ Requires libvips v8.6.1.
 * Improve install time error messages for FreeBSD users.
   [#1310](https://github.com/lovell/sharp/issues/1310)
 
+* Ensure extractChannel works with 16-bit images.
+  [#1330](https://github.com/lovell/sharp/issues/1330)
+
 * Add experimental entropy field to stats response.
 
 #### v0.20.5 - 27<sup>th</sup> June 2018